+++ DrivingLicenseAppAPI/src/main/java/csapat/DrivingLicenseAppAPI/repository/DrivingLessonRepository.java
@@ -11,20 +11,20 @@ import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
-public interface DrivingLessonRepository extends JpaRepository<DrivingLessons, Integer> {
+public interface DrivingLessonRepository extends JpaRepository<DrivingLessons, Long> {
 
     @Procedure(name = "getAllDrivingLesson", procedureName = "getAllDrivingLesson")
     List<DrivingLessons> getAllDrivingLesson();
 
     @Procedure(name = "getDrivingLesson", procedureName = "getDrivingLesson")
-    Optional<DrivingLessons> getDrivingLesson(@Param("idIN") Integer id);
+    Optional<DrivingLessons> getDrivingLesson(@Param("idIN") Long id);
 
     @Procedure(name = "deleteDrivingLesson", procedureName = "deleteDrivingLesson")
-    void deleteDrivingLesson(@Param("idIN") Integer id);
+    void deleteDrivingLesson(@Param("idIN") Long id);
 
     @Procedure(name = "getDrivingLessonBetweenHour", procedureName = "getDrivingLessonBetweenHour")
-    List<Integer> getDrivingLessonBetweenHour(@Param("dateIN") Date date, @Param("startHourIN") Date startHour, @Param("endHourIN") Date endHour, @Param("instructorIDIN") Integer instructorId);
+    List<Long> getDrivingLessonBetweenHour(@Param("dateIN") Date date, @Param("startHourIN") Date startHour, @Param("endHourIN") Date endHour, @Param("instructorIDIN") Long instructorId);
 
     @Procedure(name = "getDrivingLessonByStudentId", procedureName = "getDrivingLessonByStudentId")
-    List<DrivingLessons> getDrivingLessonByStudentId(@Param("idIN") Integer id);
+    List<DrivingLessons> getDrivingLessonByStudentId(@Param("idIN") Long id);
 }
